Updates to bombs and collisions
Placing a bomb now renders a black bomb for 2 seconds, then a red one for 1 second, then disappears.
Bombs are now placed in the center of the square the character is standing on.
The character won't "bounce" against a wall anymore, when hitting it. The character will just "move in place" infront of the wall.